refactor: CouponStrategy 의 구현체 이름을 조금 더 명시적으로 수정
