+++ src/test/java/cart/coupon/domain/AllProductsCouponStrategyTest.java
@@ -9,17 +9,15 @@ import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(ReplaceUnderscores.class)
-@DisplayName("SpecificCouponStrategy 은(는)")
-class SpecificCouponStrategyTest {
+@DisplayName("AllProductsCouponStrategy 은(는)")
+class AllProductsCouponStrategyTest {
 
     @Test
-    void 특정_상품에_적용된다() {
+    void 무조건_적용된다() {
         // given
-        Long productId = 1L;
-        CouponStrategy couponStrategy = new SpecificCouponStrategy(productId);
+        CouponStrategy couponStrategy = new AllProductsCouponStrategy();
 
         // when & then
-        assertThat(couponStrategy.canApply(productId)).isTrue();
-        assertThat(couponStrategy.canApply(2L)).isFalse();
+        assertThat(couponStrategy.canApply(1L)).isTrue();
     }
 }
