+++ src/test/java/cart/coupon/domain/SpecificProductCouponStrategyTest.java
@@ -9,15 +9,17 @@ import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(ReplaceUnderscores.class)
-@DisplayName("GeneralCouponStrategy 은(는)")
-class GeneralCouponStrategyTest {
+@DisplayName("SpecificProductCouponStrategy 은(는)")
+class SpecificProductCouponStrategyTest {
 
     @Test
-    void 무조건_적용된다() {
+    void 특정_상품에_적용된다() {
         // given
-        CouponStrategy couponStrategy = new GeneralCouponStrategy();
+        Long productId = 1L;
+        CouponStrategy couponStrategy = new SpecificProductCouponStrategy(productId);
 
         // when & then
-        assertThat(couponStrategy.canApply(1L)).isTrue();
+        assertThat(couponStrategy.canApply(productId)).isTrue();
+        assertThat(couponStrategy.canApply(2L)).isFalse();
     }
 }
